working on Caribou 1.3 analysis

+++ scripts/Caribou/CM1_3_03_hexagonize_all.py
@@ -6,6 +6,8 @@
 # - loop through the mu_list
 
 
+# note that this script does not overwrite. You can run it on top of existing gdb and it will create only the fcs that is not already created.
+
 import arcpy
 import os
 
@@ -27,15 +29,26 @@ def hexagonizer(hex_base_gdb,out_gdb,join_info,mu_list,express_run):
 
 		# make a new feature dataset with MNR Lambert Conformal Conic Projection
 		DS_name = 'f_%s'%muno_txt
-		logger.print2(" -- Generating dataset: %s..."%DS_name)
-		# find the projection file location
-		projfile = os.path.join(os.path.split(os.path.split(__file__)[0])[0],"MNRLambert_d.prj")
-		if not os.path.isfile(projfile):
-			# if the file doesn't exist, show error
-			arcpy.AddError("Can't find MNRLambert_d.prj file at the parent folder of %s"%__file__)
-			raise
-		arcpy.Delete_management(DS_name)
-		arcpy.CreateFeatureDataset_management(out_gdb, DS_name, projfile)
+		DS_fullpath = os.path.join(out_gdb,DS_name)
+		DS_exists = False
+		try:
+			if arcpy.Describe(DS_fullpath).dataType.lower() == "featuredataset": DS_exists = True
+		except:
+			# will give OSError if the path doesn't exist.
+			pass
+
+		if DS_exists:
+			logger.print2(" -- %s (feature dataset) already exists"%DS_name)
+		else:
+			logger.print2(" -- Generating dataset: %s..."%DS_name)
+			# find the projection file location
+			projfile = os.path.join(os.path.split(os.path.split(__file__)[0])[0],"MNRLambert_d.prj")
+			if not os.path.isfile(projfile):
+				# if the file doesn't exist, show error
+				arcpy.AddError("Can't find MNRLambert_d.prj file at the parent folder of %s"%__file__)
+				raise
+			arcpy.Delete_management(DS_name)
+			arcpy.CreateFeatureDataset_management(out_gdb, DS_name, projfile)
 
 		# loop through the join_info dictionary
 		for layer_type, info in join_info.items():
@@ -44,39 +57,45 @@ def hexagonizer(hex_base_gdb,out_gdb,join_info,mu_list,express_run):
 			join_sql = info['sql']
 			join_include_field_list = info['keepf']
 			output_suffix = info['suffix']
-			logger.print2(" -- joining [%s]"%layer_type)
 
-
-			# make feature layer with only the join_include_field_list visible. (cause deleting them later takes way too long)
-			logger.print2("\tCreating a temporary layer")
-			fields= arcpy.ListFields(join_full_path)
-			fieldinfo = arcpy.FieldInfo()
-			for field in fields:
-				if field.name.upper() in join_include_field_list:
-					fieldinfo.addField(field.name, field.name, "VISIBLE", "NONE")
-				else:
-					fieldinfo.addField(field.name, field.name, "HIDDEN", "NONE")
-			# if there is an sql
-			if join_sql != None:
-				logger.print2("\tSelect from %s WHERE %s"%(layer_type,join_sql))				
-				arcpy.management.MakeFeatureLayer(join_full_path, "join_lyr", join_sql, field_info=fieldinfo)
-				num_selected = int(arcpy.management.GetCount("join_lyr")[0])
-				logger.print2("\t%s records selected"%num_selected)
-			else:
-				arcpy.management.MakeFeatureLayer(join_full_path, "join_lyr", field_info=fieldinfo)
-
-			# run spatial join
-			target_fc = 'f_%s_base'%muno_txt # eg. f_443_base found in hex_base_gdb
+			# prepare output fc name and path (will be used later during spatial join)
 			out_fc_name = 'f_%s_%s'%(muno_txt,output_suffix)
 			out_fc_fullpath = os.path.join(out_gdb,DS_name,out_fc_name)
-			logger.print2("\tRunning Spatial Join. out_fc_name = %s"%out_fc_name)
-			arcpy.analysis.SpatialJoin(target_features=target_fc, join_features="join_lyr", out_feature_class=out_fc_fullpath, join_operation="JOIN_ONE_TO_ONE", join_type="KEEP_COMMON", match_option="HAVE_THEIR_CENTER_IN")
 
-			# delete extra fields
-			if not express_run:
-				delete_field_lst = ['SHAPE_AREA_1','SHAPE_LENGTH_1']
-				logger.print2("\tClening up - deleting %s"%delete_field_lst)
-				arcpy.management.DeleteField(out_fc_fullpath, delete_field_lst)
+			# only run if the output doesn't already exists.
+			outfc_already_exists = arcpy.Exists(out_fc_fullpath)
+			if outfc_already_exists:
+				logger.print2(" -- %s already exists"%out_fc_name)
+			else:
+				# make feature layer with only the join_include_field_list visible. (cause deleting them later takes way too long)
+				logger.print2(" -- joining [%s]"%layer_type)
+				logger.print2("\tCreating a temporary layer")
+				fields= arcpy.ListFields(join_full_path)
+				fieldinfo = arcpy.FieldInfo()
+				for field in fields:
+					if field.name.upper() in join_include_field_list:
+						fieldinfo.addField(field.name, field.name, "VISIBLE", "NONE")
+					else:
+						fieldinfo.addField(field.name, field.name, "HIDDEN", "NONE")
+				# if there is an sql
+				if join_sql != None:
+					logger.print2("\tSelect from %s WHERE %s"%(layer_type,join_sql))				
+					arcpy.management.MakeFeatureLayer(join_full_path, "join_lyr", join_sql, field_info=fieldinfo)
+					num_selected = int(arcpy.management.GetCount("join_lyr")[0])
+					logger.print2("\t%s records selected"%num_selected)
+				else:
+					arcpy.management.MakeFeatureLayer(join_full_path, "join_lyr", field_info=fieldinfo)
+
+				# run spatial join
+				target_fc = 'f_%s_base'%muno_txt # eg. f_443_base found in hex_base_gdb
+				logger.print2("\tRunning Spatial Join. out_fc_name = %s"%out_fc_name)
+				arcpy.analysis.SpatialJoin(target_features=target_fc, join_features="join_lyr", out_feature_class=out_fc_fullpath, join_operation="JOIN_ONE_TO_ONE", join_type="KEEP_COMMON", match_option="HAVE_THEIR_CENTER_IN")
+
+				# delete extra fields
+				if not express_run:
+					delete_field_lst = ['SHAPE_AREA_1','SHAPE_LENGTH_1']
+					logger.print2("\tClening up - deleting %s"%delete_field_lst)
+					arcpy.management.DeleteField(out_fc_fullpath, delete_field_lst)
 
 
 
@@ -92,6 +111,7 @@ if __name__ == '__main__':
 	hex_base_gdb = r'C:\Users\KimDan\Government of Ontario\Caribou Explorer - CM1_3\Data\Analysis\Hexagon_Test\HexBase.gdb' # will have layers such as 'f_443' and 'f_816'
 	out_gdb = r'C:\Users\KimDan\Government of Ontario\Caribou Explorer - CM1_3\Data\Analysis\Hexagon_Test\HexJoin.gdb' # must already exist
 	ARAR_path = r'C:\Users\KimDan\Government of Ontario\Forest Explorer - Data\D\AR\AnalysisReadyAR.gdb'
+	ARI_path = r'C:\Users\KimDan\Government of Ontario\Forest Explorer - FRO\FRO2026\02InventoryCleanUp\ARI_wFU.gdb'
 
 	join_info = {
 	'EventLayer': 	{'path':r'C:\Users\KimDan\Government of Ontario\Caribou Explorer - CM1_3\Data\Analysis\EventLayers\Event2002to2023.gdb\NEO4\NEO_fin',
@@ -113,12 +133,11 @@ if __name__ == '__main__':
 	 				'suffix':'range',
 	 				},	 	
 
-	# adding spcomp95 takes too long - may be I need to compartmentalize spcomp95 first before doing spatial join by adding MUNO field and filling that out first.
-	# 'spcomp95':		{'path': r'C:\Users\KimDan\Government of Ontario\Forest Explorer - Data\D\GeoData\SpComp95\SpComp95.gdb\SpComp95', 
-	#  				'sql': None,
-	#  				'keepf': ['SPCOMP','LG_Region'],
-	#  				'suffix':'spc95',
-	#  				},
+	'spcomp95':		{'path': r'C:\Users\KimDan\Government of Ontario\Caribou Explorer - CM1_3\Data\Analysis\SpComp95\SPCOMP95_mod_v2026.gdb\SPCOMP95_wSFU', 
+	 				'sql': None,
+	 				'keepf': ['SPCOMP','SFU_REGION','ECONUM','SFU','PFT','LGFU'],
+	 				'suffix':'spc95',
+	 				},
 
 	'Harvest':		{'path': os.path.join(ARAR_path,'Harvest','HRV_All_02_n_up'),
 	 				'sql': None,
@@ -155,9 +174,17 @@ if __name__ == '__main__':
 	 				'keepf': ['TRTMTHD1', 'TRTCAT1', 'PRODTYPE', 'RATE_AI', 'APPNUM', 'AR_YEAR'],
 	 				'suffix':'sip',
 	 				},
+
+	'ARI25':		{'path': os.path.join(ARI_path,'FC$MUNO$'),
+	 				'sql': "POLYTYPE = 'FOR'",
+	 				'keepf': ['YRSOURCE','SOURCE','DEVSTAGE','YRDEP','SILVSYS','DEPTYPE','YRORG','LEADSPC','SPCOMP','AGE','HT','CCLO','STKG','SC','PLANFU','SGR','SFU','LGFU','PFT'],
+	 				'suffix':'ARI25',
+	 				},	 				
 	}
 
-	# add latest ARI
+	# add 2011 FRI T1
+	# add KMB stuff as they come in
+	# add RAP?
 
 
 
@@ -167,7 +194,7 @@ if __name__ == '__main__':
 			'FC780', 'FC796', 'FC816', 'FC840', 'FC889', 'FC898', 'FC930', 'FC966', 'FC994'] # 39 MUs
 	mu_list = ['FC443','FC816']
 
-	express_run = True
+	express_run = False
